Solution4

+++ app/main.py
@@ -1,13 +1,21 @@
+from typing import Type, Optional
+from types import TracebackType
 import os
 
 
 class CleanUpFile:
 
-    def __init__(self, filename) -> None:
+    def __init__(self, filename: str) -> None:
         self.filename = filename
 
-    def __enter__(self) -> object:
+    def __enter__(self) -> "CleanUpFile":
         return self
 
-    def __exit__(self, exc_type, exc, tb) -> None:
+    def __exit__(
+        self,
+        exc_type: Optional[Type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[TracebackType]
+    ) -> None:
+        if os.path.exists(self.filename):
             os.remove(self.filename)
